backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
langchain==0.1.11
openai==1.12.0
pypdf==4.1.0
python-docx==1.1.0
chromadb==0.4.24
werkzeug==3.0.1
gunicorn==21.2.0